Improve determinstic hashes test
Originally, we were just testing to see that the hash was the same twice
in a row. This test often passed when our hashes were not deterministic.
I want to make this test more likely to fail when hashes are not
deterministic, so I am running it 20 times. This should help us ensure
that these hashes remain deterministic.