Default to branch named "main"
In cases where we can't figure out a baseline commit in our CI scripts,
we default to `origin/master`. This was the usual default branch name
back in the days, but it has since changed to `origin/main`. I'm making
the switch to "main" now.

This is a breaking change but one that is easy to make non-breaking if
you use a BASE_BRANCH=origin/master environment variable.