Refactor archiving to go through single determinstic function
We had three places that were interfacing with the archiver package, and
they all need to do basically the same thing--create a deterministic zip
file. If the zip file is nondeterministic, then downstream problems with
caching will arise. Since creating a deterministic zip requires some
special handling, I decided that it would be best to refactor this code
to all run through the same function that guarantees a deterministic
output.

While working on this I noticed that the remoteRunner was using
archiver's directory function, which I believe can by nondeterministic,
so this should help to stabilize those zip files.