Fix deterministicArchive for relative paths
When I tried out v12.0.0-beta.0, I found that it broke the Storybook
plugin. I noticed that the plugin provides a relative path to be zipped,
and in the new version the files that were extracted from this zip were
deeply nested in the directory structure they were relative to, instead
of being at the top level of the zip.

To fix this, we need to resolve the dir before slicing it, so that we
slice up to the dir that contains the files.

I was able to reproduce this in the tests and I added a test.

I also decided to remove some paths to make it easier to understand
which paths are used for what. archiver.file just delegates the work to
append internally anyway, so this should not be a problem at all.