Add way to opt in to uploading assets with signed URLs
This will allow folks to opt in to using signed URLs by setting the
HAPPO_SIGNED_URL env var.

This requires a separate change to land on the server side before it can
be used.

To support this change, I ended up doing some bigger refactoring to the
react integration test. This change should make the test a lot more
realistic, because it now avoids mocking the makeRequest method in favor
of spinning up a server that responds with mock responses.