Keep dir structure for single files in deterministicArchive
I found an issue with happo-cypress and inlined canvas images.
Basically, the canvas element is converted to an image and inserted into
the page, plus added to the asset package for the happo-cypress run. In
old versions of the asset packages, the inlined canvas image would be
kept in a `.happo-tmp/inlined/<hash>.png` file. In recent versions, the
file is instead stored directly in the root of the asset package,
`./<hash>.png`.

In happo-e2e, this commit introduce the bug to happo-cypress:
https://github.com/happo/happo-e2e/commit/1cfe6ffbcb88ca638728918226dfbb632cc97866
(Use deterministicArchive.js from main happo.io library)

By keeping the directory structure, these entries will be stored in the
right folder, so that the image elements will reference the right thing:

<img src=".happo-tmp/_inlined/<hash>.png">