Add support for HAPPO_GITHUB_TOKEN when posting comment
This is the GitHub Enterprise scenario where instead of a status post we
post a comment to the PR with a status image and a link to the Happo
comparison report.

Our legacy support will use a personal access token and basic auth to
post a comment. This works but it makes it weird because the comment
looks like it's coming from a user. And if that user leaves the company
they have to update the credentials etc.

When using GitHub actions, we are fed with a secrets.GITHUB_TOKEN that
has the right permissions to post a comment to the PR. It is
short-lived which means it is more secure. We can allow people to pass
it in using HAPPO_GITHUB_TOKEN and then use it instead of
HAPPO_GITHUB_USER_CREDENTIALS.