Delete old comments as part of posting new one
This is on a request from a customer. They are using comment posting and
noticed that the list of comments grow quite long on a PR, especially
when you are pushing a lot of changes to it. We can make this better by
removing all old comments before posting our new one.

To make sure that only Happo comments are removed I'm adding a marker at
the top of each comment (html comment) that we will use to make sure we
only remove relevant comments.

This new behavior is opt-in and is only active when
HAPPO_DELETE_OLD_COMMENTS="true"