Use git to determine previous SHA in GitHub actions
I've noticed that we often end up using the wrong report for comparisons
on GitHub actions. I think this is because the event JSON gives us
`pull_request.base.sha` as the tip of the base branch and not the merge
base SHA.

To make this work better, I think we want to use `git merge-base`. This
is similar to what we do in the CircleCI integration.